// yugabyted-ui/ui/src/api/src/models/XClusterOutboundGroup.ts
// tslint:disable
/**
 * Yugabyte Cloud
 * YugabyteDB as a Service
 *
 * The version of the OpenAPI document: v1
 * Contact: support@yugabyte.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// eslint-disable-next-line no-duplicate-imports
import type { TableReplicationLagDetailsOutbound } from './TableReplicationLagDetailsOutbound';


/**
 * Schema for xCluster replication group details.
 * @export
 * @interface XClusterOutboundGroup
 */
export interface XClusterOutboundGroup  {
  /**
   * Unique identifier for the replication group.
   * @type {string}
   * @memberof XClusterOutboundGroup
   */
  replication_group_id: string;
  /**
   * The current status of the replication group (e.g., \'active\', \'inactive\').
   * @type {string}
   * @memberof XClusterOutboundGroup
   */
  state: string;
  /**
   * The list of IP addresses of the nodes involved in the replication.
   * @type {string[]}
   * @memberof XClusterOutboundGroup
   */
  replication_cluster_node_ips?: string[];
  /**
   * UUID of the source universe for the replication group.
   * @type {string}
   * @memberof XClusterOutboundGroup
   */
  source_universe_uuid?: string;
  /**
   * UUID of the target universe for the replication group.
   * @type {string}
   * @memberof XClusterOutboundGroup
   */
  target_universe_uuid: string;
  /**
   * List of namespaces from outbound and lag of each table
   * @type {TableReplicationLagDetailsOutbound[]}
   * @memberof XClusterOutboundGroup
   */
  tables_list_with_lag?: TableReplicationLagDetailsOutbound[];
}