// yugabyted-ui/ui/src/api/src/models/XClusterInboundGroup.ts
// tslint:disable
/**
 * Yugabyte Cloud
 * YugabyteDB as a Service
 *
 * The version of the OpenAPI document: v1
 * Contact: support@yugabyte.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




/**
 * Schema for xCluster replication group details.
 * @export
 * @interface XClusterInboundGroup
 */
export interface XClusterInboundGroup  {
  /**
   * Unique identifier for the replication group.
   * @type {string}
   * @memberof XClusterInboundGroup
   */
  replication_group_id: string;
  /**
   * The current status of the replication group (e.g., \'active\', \'inactive\').
   * @type {string}
   * @memberof XClusterInboundGroup
   */
  state: string;
  /**
   * The list of IP addresses of the nodes involved in the replication on target side.
   * @type {string[]}
   * @memberof XClusterInboundGroup
   */
  target_cluster_node_ips?: string[];
  /**
   * The list of IP addresses of the nodes involved in the replication on source side.
   * @type {string[]}
   * @memberof XClusterInboundGroup
   */
  source_cluster_node_ips?: string[];
  /**
   * UUID of the source universe for the replication group.
   * @type {string}
   * @memberof XClusterInboundGroup
   */
  source_universe_uuid?: string;
  /**
   * UUID of the target universe for the replication group.
   * @type {string}
   * @memberof XClusterInboundGroup
   */
  target_universe_uuid?: string;
}